+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## v_0.0.1（2016-06-08）
+## v_0.0.2 (2016-12-08)
+
+- (-) Remove clean command.
+- (+) Add image compression command.
+
+## v_0.0.1 (2016-06-08)
 
 - (+) Add clean command.